Add IdAwareComponentRegistry
Same as NamedComponentRegistry, but for objects implementing IdAwareObject. Simplifies the logic for getKey and removes the name specific register method compared to the NamedComponentRegistry
